Add no apps error
