+++ components/faphub/installedtab/impl/src/main/res/values/strings.xml
@@ -4,4 +4,5 @@
     <string name="faphub_installed_update_cancel">CANCEL UPDATING</string>
     <string name="faphub_installed_offline">Connect to the internet to see available updates</string>
     <string name="faphub_installed_offline_btn">INSTALLED</string>
+    <string name="faphub_installed_empty_desc">No apps installed on your Flipper yet</string>
 </resources>
