+++ components/faphub/uninstallbutton/api/build.gradle.kts
