+++ components/faphub/uninstallbutton/impl/src/main/kotlin/com/flipperdevices/faphub/uninstallbutton/impl/viewmodel/DeleteViewModel.kt
